+++ lgtv-drivers/CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.4.0] - 2025-01-23
+### Added
+- Add "Take" command to take a TV screenshot
+
+### Fixed
+- Correctly identify event type (digital vs physical)
+
+### Fixed
+- Hide errors for unsupported config keys - `@nclark`
+
 ## [1.3.0] - 2025-01-23
 ### Added
 - Add "Screen Saver On" and "Screen Saver Off" commands